zombie hitbox not enabled by animation
Make it so that a zombie's hitbox for his attack is not
enabled/disabled using the animation curve system. Instead
the hitbox is enabled/disbaled using events which are triggered
in the appropriate animation frame. This way a hitbox is only
enabled once, and will nto remain enabled if the hitbox is disabled
during the remainder of the animation.